fix(be/spdk): don't detach in-use enumerated controllers
If XNVME_ENUMERATE_DEV_KEEP_OPEN is returned from the callback during
enumeration, the SPDK controller underlying the open device must not be
detached. Only deref the controller if the device is closed.
Conditionally detach controllers only if their reference count is zero.

It is now valid for a controller to have a reference count of zero and
not be cleaned up yet. Entries in the reference count array may not be
reused if the controller pointer is non-null. A controller with a zero
reference count may be still be incremented later. A new cleanup routine
allows deferred detachment of all controllers with a zero refcount.
Insert and lookup are combined into a single function, removing the
danger of accidentally inserting the same controller twice.

Tests are added to verify keeping devices open during enumeration.

Signed-off-by: Alex Docauer <alex@docauer.net>
Co-authored-by: Karl Bonde Torp <karlowich@gmail.com>